Change namespaces at data layer + make socket logic more easier
