+++ OnlinerTask/OnlinerTask.Data/Hubs/PersonalHub.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNet.SignalR;
-using OnlinerTask.Data.RedisManager.RedisServer.RedisRequests;
+using OnlinerTask.Data.RedisManager.RedisServer.RedisRequests.Implementations;
 
 namespace OnlinerTask.Data.Hubs
 {
