+++ OnlinerTask/OnlinerTask.Data/RedisManager/RedisServer/RedisRequests/Implementations/RemoveProductRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace OnlinerTask.Data.RedisManager.RedisServer.RedisRequests
+﻿namespace OnlinerTask.Data.RedisManager.RedisServer.RedisRequests.Implementations
 {
     public class RemoveProductRequest : INotifyRequest
     {
